+++ R/lua_internal.R
@@ -1,0 +1,3 @@
+# Access to Lua API
+
+lua_gettop = function(L = NULL) .Call(`_luajr_lua_gettop`, L)