+++ inst/include/luajr_const.h
@@ -1,0 +1,7 @@
+#ifndef LUAJR_CONST_H
+#define LUAJR_CONST_H
+
+#define LUAJR_TOOLING_NONE 0
+#define LUAJR_TOOLING_ALL 1
+
+#endif